test(gt): keep fixtures local for now to keep PR scoped
